+++ docs/lang/articles/c-api/taichi_core.md
@@ -10,14 +10,14 @@ Taichi Core exposes all necessary interfaces for offloading the AOT modules to T
 
 Taichi C-API intends to support the following backends:
 
-|Backend|Offload Target|Maintenance Tier|
-|-|-|-|
-|Vulkan|GPU|Tier 1|
-|CUDA (LLVM)|GPU (NVIDIA)|Tier 1|
-|CPU (LLVM)|CPU|Tier 1|
-|OpenGL|GPU|Tier 2|
-|DirectX 11|GPU (Windows)|N/A|
-|Metal|GPU (macOS, iOS)|N/A|
+|Backend     |Offload Target   |Maintenance Tier |
+|------------|-----------------|-----------------|
+|Vulkan      |GPU              |Tier 1           |
+|CUDA (LLVM) |GPU (NVIDIA)     |Tier 1           |
+|CPU (LLVM)  |CPU              |Tier 1           |
+|OpenGL      |GPU              |Tier 2           |
+|DirectX 11  |GPU (Windows)    |N/A              |
+|Metal       |GPU (macOS, iOS) |N/A              |
 
 The backends with tier-1 support are being developed and tested more intensively. And most new features will be available on Vulkan first because it has the most outstanding cross-platform compatibility among all the tier-1 backends.
 For the backends with tier-2 support, you should expect a delay in the fixes to minor issues.
@@ -671,6 +671,7 @@ typedef enum TiFormat {
   TI_FORMAT_MAX_ENUM = 0xffffffff,
 } TiFormat;
 ```
+
 ---
 ### Structure `TiImageOffset`
 
@@ -766,6 +767,7 @@ typedef enum TiFilter {
   TI_FILTER_MAX_ENUM = 0xffffffff,
 } TiFilter;
 ```
+
 ---
 ### Enumeration `TiAddressMode`
 
@@ -778,6 +780,7 @@ typedef enum TiAddressMode {
   TI_ADDRESS_MODE_MAX_ENUM = 0xffffffff,
 } TiAddressMode;
 ```
+
 ---
 ### Structure `TiSamplerCreateInfo`
 
@@ -790,6 +793,7 @@ typedef struct TiSamplerCreateInfo {
   float max_anisotropy;
 } TiSamplerCreateInfo;
 ```
+
 ---
 ### Structure `TiTexture`
 
@@ -1008,6 +1012,7 @@ TI_DLL_EXPORT TiSampler TI_API_CALL ti_create_sampler(
   const TiSamplerCreateInfo* create_info
 );
 ```
+
 ---
 ### Function `ti_destroy_sampler`
 
@@ -1018,6 +1023,7 @@ TI_DLL_EXPORT void TI_API_CALL ti_destroy_sampler(
   TiSampler sampler
 );
 ```
+
 ---
 ### Function `ti_create_event`
