+++ misc/generate_c_api_docs.py
@@ -1,128 +1,18 @@
-import re
-from collections import defaultdict
-from pathlib import Path
-
-from generate_c_api import get_declr, get_human_readable_name
-from taichi_json import (Alias, BitField, BuiltInType, Definition, EntryBase,
-                         Enumeration, Field, Function, Handle, Module,
-                         Structure, Union)
-
-SYM_PATTERN = r"\`(\w+\.\w+(?:\.\w+)?)\`"
+from generate_c_api import (get_declr, get_title,
+                            resolve_inline_symbols_to_names)
+from taichi_json import BuiltInType, Module
 
 
-def get_title(x: EntryBase):
-    if isinstance(x, BuiltInType):
-        return ""
-
-    extra = ""
-    if isinstance(x, Function) and x.is_device_command:
-        extra += " (Device Command)"
-
-    if isinstance(x, (Alias, Definition, Handle, Enumeration, BitField,
-                      Structure, Union, Function)):
-        return f"{type(x).__name__} `{get_human_readable_name(x)}`" + extra
-    else:
-        raise RuntimeError(f"'{x.id}' doesn't need title")
-
-
-def get_human_readable_field_name(x: EntryBase, field_name: str):
-    out = None
-    if isinstance(x, Enumeration):
-        out = x.name.extend(field_name).screaming_snake_case
-    elif isinstance(x, BitField):
-        out = x.name.extend(field_name).extend('bit').screaming_snake_case
-    elif isinstance(x, Structure):
-        for field in x.fields:
-            if str(field.name) == field_name:
-                out = str(field.name)
-                break
-    elif isinstance(x, Union):
-        for field in x.variants:
-            if str(field.name) == field_name:
-                out = str(field.name)
-                break
-    elif isinstance(x, Function):
-        for field in x.params:
-            if str(field.name) == field_name:
-                out = str(field.name)
-                break
-    return out
-
-
-def resolve_symbol_to_name(module: Module, id: str):
-    """Returns the resolved symbol and its hyperlink (if available)"""
-    try:
-        ifirst_dot = id.index('.')
-    except ValueError:
-        return None
-
-    field_name = ""
-    try:
-        isecond_dot = id.index('.', ifirst_dot + 1)
-        field_name = id[isecond_dot + 1:]
-        id = id[:isecond_dot]
-    except ValueError:
-        pass
-
-    out = module.declr_reg.resolve(id)
-    href = None
-
-    try:
-        if field_name:
-            out = get_human_readable_field_name(out, field_name)
-        else:
-            href = "#" + get_title(out).lower().replace(' ', '-').replace(
-                '`', '').replace('(', '').replace(')', '')
-            out = get_human_readable_name(out)
-    except:
-        print(f"WARNING: Unable to resolve symbol {id}")
-        out = id
-
-    return out, href
-
-
-def resolve_inline_symbols_to_names(module: Module, line: str):
-    matches = re.findall(SYM_PATTERN, line)
-
-    replacements = {}
-    for m in matches:
-        id = str(m)
-        replacements[id] = resolve_symbol_to_name(module, id)
-
-    for old, (new, href) in replacements.items():
-        if new is None:
-            print(f"WARNING: Unresolved inline symbol `{old}`")
-        else:
-            if href is None:
-                new = f"`{new}`"
-            else:
-                new = f"[`{new}`]({href})"
-            line = line.replace(f"`{old}`", new)
-    return line
-
-
-def print_module_doc(module: Module, templ):
-    out = []
-
-    for i in range(len(templ)):
-        line = templ[i].strip()
-        line = resolve_inline_symbols_to_names(module, line)
-        out += [line]
-        if line.startswith("## API Reference"):
-            break
-
+def print_module_doc(module: Module):
+    out = ["---"]
+    out += module.doc.markdown_metadata
+    out += [
+        "---",
+        ""
+    ]
+    for line in module.doc.module_doc:
+        out += [resolve_inline_symbols_to_names(module, line)]
     out += [""]
-
-    cur_sym = None
-    documented_syms = defaultdict(list)
-    for line in templ[i:]:
-        line = line.strip()
-        if re.match(SYM_PATTERN, line):
-            cur_sym = line[1:-1]
-            continue
-        documented_syms[cur_sym] += [
-            resolve_inline_symbols_to_names(module, line)
-        ]
 
     is_first = True
     for x in module.declr_reg:
@@ -142,35 +32,24 @@ def print_module_doc(module: Module, templ):
             "```",
         ]
 
-        if x in documented_syms:
-            out += documented_syms[x]
+        if x in module.doc.api_refs:
+            for line in module.doc.api_refs[x]:
+                out += [resolve_inline_symbols_to_names(module, line)]
         else:
             print(f"WARNING: `{x}` is not documented")
-
-    out += [""]
+        out += [""]
 
     return '\n'.join(out)
 
 
 def generate_module_header(module):
-    if module.is_built_in:
-        return
-
-    templ_path = f"c_api/docs/{module.name}.md"
-    templ = None
-    if Path(templ_path).exists():
-        with open(templ_path) as f:
-            templ = f.readlines()
-    else:
-        print(
-            f"ignored {templ_path} because the documentation template cannot be found"
-        )
+    if module.is_built_in or module.doc is None:
         return
 
     print(f"processing module '{module.name}'")
     path = f"docs/lang/articles/c-api/{module.name[7:-2]}.md"
     with open(path, "w") as f:
-        f.write(print_module_doc(module, templ))
+        f.write(print_module_doc(module))
 
     #system(f"clang-format {path} -i")
 
@@ -204,6 +83,7 @@ if __name__ == "__main__":
         BuiltInType("VkImageViewType", "VkImageViewType"),
         BuiltInType("PFN_vkGetInstanceProcAddr", "PFN_vkGetInstanceProcAddr"),
         BuiltInType("char", "char"),
+        BuiltInType("GLuint", "GLuint"),
     }
 
     for module in Module.load_all(builtin_tys):
